Enrollment Dashboard
This dashboard was developed about 2 years ago. It is using dummy data.  A lot of information is also added as a tooltip, which can be accessed by hovering over the visuals. A data dictionary is also added to the dashboard, which can be accessed by hovering over the bulb icon on the top right of the dashboard.